pyproject.toml
[tool.poetry]
name = "franky_gendiff"
version = "0.6.3"
authors = ["Roman Kornveyts <fcdm@inbox.ru>"]
description = ""

packages = [
  { include = "gendiff" },
]

[tool.poetry.dependencies]
python = "^3.8"
PyYAML = "^5.3"

[tool.poetry.dev-dependencies]
flake8 = "^3.7.9"
pytest = "^5.3.5"
pytest-cov = "^2.8.1"


[tool.poetry.scripts]
gendiff = "gendiff.scripts:main"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"